agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
